pyproject.toml
[tool.poetry]
name = "mainsitebackend"
version = "0.1.0"
description = ""
authors = ["art3mis3d <achuqw@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
django-filter = "^23.2"
django-split-settings = "^1.2.0"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[tool.isort]
multi_line_output = 5
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
